run, virtme-ng-init: accept `--login` to start a login shell
I prefer having a login shell inside virtme-ng because several elements
of the work environment I'm used to (like tmux) are all tied to the
login shell initialization.

In addition to that, cleanup user command handling generally.
Use `bash -l` if we aren't changing the user, and `su -l` if we are.
While we're at it, stop unnecessarily chaining shells and run su(1)
directly if we have to.

Signed-off-by: Ivan Shapovalov <intelfx@intelfx.name>