run: explicitly specify `format=raw` for disk images
Not specifying a disk image file format leads to QEMU issuing a warning
and write-protecting block 0. This is seldom desirable:

```
WARNING: Image format was not specified for '/path/to/file' and probing guessed raw.
         Automatically detecting the format is dangerous for raw images, write operations on block 0 will be restricted.
         Specify the 'raw' format explicitly to remove the restrictions.
```

To squelch the warning, specify `format=raw` explicitly. This will be
made configurable in the next commit.

Signed-off-by: Ivan Shapovalov <intelfx@intelfx.name>