vng, run: truncate scsi device_id if the full serial does not fit
The scsi-hd `serial=` parameter is used by QEMU as a default for the
`device_id=` parameter (which is what in fact ends up in the by-id
path).

The `serial=` parameter is set by `virtme-run` using the disk "name"
provided in the `--{,blk-}disk` argument, which is in turn set by `vng`
to the disk path, when not explicitly provided.

When all of this is chained, the full path to disk image ends up being
used as the SCSI device_id:

1. `vng --disk /path/to/disk` leads to
2. `virtme-run --disk /path/to/disk=/path/to/disk` leads to
3. `qemu -device scsi-hd,serial=/path/to/disk` leads to (internally)
4. `qemu -device scsi-hd,serial=/path/to/disk,device_id=/path/to/disk`

However, QEMU (or perhaps it is SCSI) limits the device_id to
20 characters:

```
qemu-system-x86_64: -device scsi-hd,drive=disk0,vendor=virtme,product=disk,serial=/long/path/to/image/file: The serial number can't be longer than 20 characters when it is also used as the default for device_id
```

This problem is indirectly exposed since the previous commit, because
`vng --disk` is now a shorthand for a SCSI disk, not a virtio-blk disk.

Fix this by truncating the SCSI serial manually and using the truncated
value as the SCSI device_id.

Signed-off-by: Ivan Shapovalov <intelfx@intelfx.name>